move camera into cameraDevice file/capability
