+++ packages/teams-js/src/public/audioVisualDevice.ts
@@ -1,24 +1,17 @@
 /* eslint-disable @typescript-eslint/explicit-member-accessibility */
 
-import { sendAndHandleSdkError, sendMessageToParent, sendMessageToParentAsync } from '../internal/communication';
+import { sendAndHandleSdkError, sendMessageToParent } from '../internal/communication';
 import { getMediaCallbackSupportVersion, mediaAPISupportVersion } from '../internal/constants';
 import { registerHandler, removeHandler } from '../internal/handlers';
 import { ensureInitialized, isCurrentSDKVersionAtLeast } from '../internal/internalAPIs';
-import {
-  createFile,
-  decodeAttachment,
-  throwExceptionIfMediaCallIsNotSupportedOnMobile,
-  validateGetMediaInputs,
-  validateSelectMediaInputs,
-  validateViewImagesInput,
-} from '../internal/mediaUtil';
+import { createFile, decodeAttachment, validateGetMediaInputs } from '../internal/mediaUtil';
 import {
   callCallbackWithErrorOrResultFromPromiseAndReturnPromise,
   generateGUID,
   InputFunction,
 } from '../internal/utils';
-import { FrameContexts, MediaType } from './constants';
-import { ErrorCode, ImageProps, SdkError } from './interfaces';
+import { FrameContexts } from './constants';
+import { ErrorCode, SdkError } from './interfaces';
 import { media } from './media';
 import { runtime } from './runtime';
 
@@ -169,87 +162,5 @@ export namespace audioVisualDevice {
 
   export function isSupported(): boolean {
     return runtime.supports.media ? true : false;
-  }
-
-  export namespace camera {
-    export interface ImageInputs {
-      mediaType: MediaType.Image;
-      maxMediaCount: number;
-      imageProps?: ImageProps;
-    }
-
-    /**
-     * Input for view images API
-     */
-    export interface ImageUri {
-      value: string;
-      type: ImageUriType;
-    }
-
-    /**
-     * ID contains a mapping for content uri on platform's side, URL is generic
-     */
-    export enum ImageUriType {
-      ID = 1,
-      URL = 2,
-    }
-
-    export function selectImages(imageInputs: ImageInputs): Promise<media.Media[]> {
-      ensureInitialized(FrameContexts.content, FrameContexts.task);
-
-      // Probably should clean this up, no reason to use this structure anymore
-      const wrappedFunction: InputFunction<media.Media[]> = () =>
-        new Promise<[SdkError, media.Media[]]>(resolve => {
-          if (!isCurrentSDKVersionAtLeast(mediaAPISupportVersion)) {
-            throw { errorCode: ErrorCode.OLD_PLATFORM };
-          }
-          throwExceptionIfMediaCallIsNotSupportedOnMobile(imageInputs);
-
-          if (!validateSelectMediaInputs(imageInputs)) {
-            throw { errorCode: ErrorCode.INVALID_ARGUMENTS };
-          }
-
-          const params = [imageInputs];
-          // What comes back from native at attachments would just be objects and will be missing getMedia method on them.
-          resolve(sendMessageToParentAsync<[SdkError, media.Media[]]>('selectMedia', params));
-        }).then(([err, localAttachments]: [SdkError, media.Media[]]) => {
-          // Media Attachments are final response to selectMedia
-          if (!localAttachments) {
-            throw err;
-          }
-          const mediaArray: media.Media[] = [];
-          for (const attachment of localAttachments) {
-            mediaArray.push(new media.Media(attachment));
-          }
-          return mediaArray;
-        });
-
-      return callCallbackWithErrorOrResultFromPromiseAndReturnPromise<media.Media[]>(wrappedFunction);
-    }
-
-    /**
-     * View images using native image viewer
-     *
-     * @param uriList - list of URIs for images to be viewed - can be content URI or server URL. Supports up to 10 Images in a single call
-     * @returns A promise resolved when the viewing action is completed or rejected with an @see SdkError
-     */
-    export function viewImages(uriList: ImageUri[]): Promise<void> {
-      ensureInitialized(FrameContexts.content, FrameContexts.task);
-
-      return new Promise<void>(resolve => {
-        if (!isCurrentSDKVersionAtLeast(mediaAPISupportVersion)) {
-          throw { errorCode: ErrorCode.OLD_PLATFORM };
-        }
-        if (!validateViewImagesInput(uriList)) {
-          throw { errorCode: ErrorCode.INVALID_ARGUMENTS };
-        }
-
-        resolve(sendAndHandleSdkError('viewImages', uriList));
-      });
-    }
-
-    export function isSupported(): boolean {
-      return runtime.supports.media.camera ? true : false;
-    }
   }
 }
